[add] - add implementation of services and major update comming with it
