--- genesis-core/src/main/resources/data_genesis/frontend/skeletoons/vue/vue-3-ts-skeletoon/src/core/button/GenesisAddButtonComponent.ts
@@ -1,13 +1,0 @@
-import GenesisButton from "./GenesisButton.vue";
-import { defineComponent } from "vue";
-
-export default defineComponent({
-  name: "GenesisAddButton",
-  components: { GenesisButton },
-  props: {
-    label: {
-      required: false,
-      type: String,
-    },
-  },
-});
