--- genesis-core/src/main/resources/data_genesis/frontend/skeletoons/vue/vue-3-ts-skeletoon/src/core/form/GenesisInputComponent.ts
@@ -1,15 +1,0 @@
-import { computed, defineComponent } from "vue";
-
-export default defineComponent({
-  props: {
-    label: { type: String, required: false },
-    modelValue: { type: [String, Number, Date], required: false },
-  },
-  setup(props) {
-    const inputId = computed(() => {
-      return props.label ? "inpt" + props.label : "";
-    });
-
-    return { inputId };
-  },
-});
