--- genesis-core/src/main/resources/data_genesis/frontend/skeletoons/vue/vue-3-ts-skeletoon/src/core/navbar/GenesisSidebarComponent.ts
@@ -1,10 +1,0 @@
-import { defineComponent } from "vue";
-import UserNavProfile from "../user/UserNavProfile.vue";
-
-export default defineComponent({
-  name: "GenesisSidebarComponent",
-  components: { UserNavProfile },
-  setup() {
-    return;
-  },
-});
