--- genesis-core/src/main/resources/data_genesis/frontend/skeletoons/vue/vue-3-ts-skeletoon/src/core/user/UserNavProfileComponent.ts
@@ -1,24 +1,0 @@
-import { computed, defineComponent } from "vue";
-import GenesisButton from "../button/GenesisButton.vue";
-
-export default defineComponent({
-  name: "UserNavProfile",
-  props: {
-    name: {
-      required: true,
-      type: String,
-    },
-  },
-  components: { GenesisButton },
-  setup(props) {
-    const initials = computed(() => {
-      if (!props.name) return "";
-      return props.name
-        .split(" ")
-        .map((n) => n[0])
-        .join("")
-        .toUpperCase();
-    });
-    return { initials };
-  },
-});
