--- genesis-core/src/main/resources/data_genesis/frontend/skeletoons/vue/vue-3-ts-skeletoon/src/models/DepartementsModel.ts
@@ -1,17 +1,0 @@
-export interface IDepartements {
-  id: number | undefined;
-  nomDepartement: string | undefined;
-  codeDepartement: string | undefined;
-}
-
-export class Departements implements IDepartements {
-  id: number | undefined;
-  nomDepartement: string | undefined;
-  codeDepartement: string | undefined;
-
-  constructor(data?: Partial<IDepartements>) {
-    this.id = data?.id;
-    this.nomDepartement = data?.nomDepartement;
-    this.codeDepartement = data?.codeDepartement;
-  }
-}
