+++ genesis-core/src/main/resources/data_genesis/frontend/skeletoons/vue/vue-3-ts-skeletoon/src/models/api/PageResponseModel.ts
@@ -1,30 +1,11 @@
 export interface IPageResponse<T> {
   content: Array<T>;
-  page: PageData;
+  page: IPageModel;
 }
 
-export class PageData {
-  size: number;
-  number: number;
-  totalElements: number;
-  totalPages: number;
-
-  constructor();
-  constructor(
-    size: number,
-    number: number,
-    totalElements: number,
-    totalPages: number
-  );
-  constructor(
-    size?: number,
-    number?: number,
-    totalElements?: number,
-    totalPages?: number
-  ) {
-    this.size = size ?? 10;
-    this.number = number ?? 0;
-    this.totalElements = totalElements ?? 1;
-    this.totalPages = totalPages ?? 1;
-  }
+export interface IPageModel {
+  size: number | undefined;
+  number: number | undefined;
+  totalElements: number | undefined;
+  totalPages: number | undefined;
 }
